// src/libs/openapi/models/DialogueDocumentDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DialogueDocumentMetadataDto } from './DialogueDocumentMetadataDto';
export type DialogueDocumentDto = {
    appId: string;
    documentId: string;
    content: string;
    metadata: DialogueDocumentMetadataDto;
};